Create Supervised Learning
Supervised learning is a fundamental method used in machine learning, where algorithms are taught using datasets that come with labels. This means that each piece of input data is matched with a known result, which helps the model understand and learn from examples. The aim of supervised learning is to accurately predict or classify new data by identifying patterns based on this learning process. https://www.globaltechcouncil.org/ai/supervised-learning/